Read full transaction history from CEPAS cards
Read purse Lc=01 for transaction log with start offset of 0x00 only
returns last 15 transactions. Added a second read starting with offset
0x0F to read the rest, combining the results.
Also modified wrapMessage to not set parameters.length after Lc since it's not
in the spec.

Read purse transaction log with number of bytes to read

Reading purse transaction log with starting offset but not passing number of bytes to read sometimes result in 6700 response.
Added number of bytes to read based on number of records contained in purse detail.