Add nickname feature, add option to squash dupes
This adds a feature that allows you to nickname cards, where the
nickname will be displayed instead of the serial number.

Also added is an option to overwrite current cards that match the
scanned card's serial number, rather than creating a new entry. If this
is turned on while there are already duplicate cards in the database, it
will only overwrite the latest-scanned card to avoid unintended data
loss.

Also adds whole foods to the agencies for refill data.